refactor(lsp): move changetracking to separate file
- Prefixed changetracking types with `vim.lsp.`
